# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "UAC API Documentation",
    "version": "1.0.0",
    "description": "用户认证和授权系统 API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "开发服务器"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/uploads": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "summary": "上传文件 [需要认证]",
        "description": "上传单个文件，支持图片自动转换为webp格式",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "image",
                "document",
                "video",
                "audio"
              ]
            },
            "description": "文件类型，默认为配置中的默认类型"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "要上传的文件"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "550e8400-e29b-41d4-a716-446655440000"
                        },
                        "type": {
                          "type": "string",
                          "example": "image"
                        },
                        "url": {
                          "type": "string",
                          "example": "http://localhost:3000/api/v1/uploads/images/550e8400-e29b-41d4-a716-446655440000"
                        },
                        "original_name": {
                          "type": "string",
                          "example": "example.jpg"
                        },
                        "size": {
                          "type": "integer",
                          "example": 1024
                        },
                        "mime_type": {
                          "type": "string",
                          "example": "image/jpeg"
                        },
                        "extension": {
                          "type": "string",
                          "example": ".webp"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 400
                    },
                    "message": {
                      "type": "string",
                      "example": "不支持的文件类型"
                    },
                    "data": {
                      "type": null,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/v1/uploads/multiple": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "summary": "上传多个文件 [需要认证]",
        "description": "同时上传多个文件",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "文件上传成功"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "filename": {
                            "type": "string"
                          },
                          "originalname": {
                            "type": "string"
                          },
                          "mimetype": {
                            "type": "string"
                          },
                          "size": {
                            "type": "integer"
                          },
                          "path": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/v1/uploads/image": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "summary": "上传图片 [需要认证]",
        "description": "上传单个图片，支持压缩和格式转换",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "image"
                ],
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "要上传的图片文件"
                  },
                  "compress": {
                    "type": "boolean",
                    "description": "是否压缩图片",
                    "default": true
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "jpeg",
                      "jpg",
                      "png",
                      "webp",
                      "gif"
                    ],
                    "description": "输出格式",
                    "default": "webp"
                  },
                  "width": {
                    "type": "integer",
                    "description": "压缩后的最大宽度",
                    "default": 1920
                  },
                  "height": {
                    "type": "integer",
                    "description": "压缩后的最大高度",
                    "default": 1920
                  },
                  "quality": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "description": "图片质量（1-100）",
                    "default": 80
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "图片上传成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "550e8400-e29b-41d4-a716-446655440000"
                        },
                        "filename": {
                          "type": "string",
                          "example": "550e8400-e29b-41d4-a716-446655440000.webp"
                        },
                        "originalname": {
                          "type": "string",
                          "example": "example.jpg"
                        },
                        "mimetype": {
                          "type": "string",
                          "example": "image/webp"
                        },
                        "size": {
                          "type": "integer",
                          "example": 102400
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "参数错误"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
    "/api/v1/uploads/images/{file_id}": {
      "get": {
        "tags": [
          "Uploads"
        ],
        "summary": "获取图片",
        "description": "获取指定图片，支持实时裁剪、格式转换和缓存",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "file_id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "文件ID"
          },
          {
            "in": "query",
            "name": "width",
            "schema": {
              "type": "integer"
            },
            "description": "图片宽度"
          },
          {
            "in": "query",
            "name": "height",
            "schema": {
              "type": "integer"
            },
            "description": "图片高度"
          },
          {
            "in": "query",
            "name": "quality",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "图片质量（1-100）"
          },
          {
            "in": "query",
            "name": "format",
            "schema": {
              "type": "string",
              "enum": [
                "jpeg",
                "jpg",
                "png",
                "webp",
                "gif"
              ]
            },
            "description": "输出格式"
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "image/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "文件不存在"
          },
          "500": {
            "description": "服务器错误"
          }
        }
      }
    },
...
  },
...
}